# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ErrorResponse:
    properties:
      error:
        example: invalid_credentials
        type: string
    type: object
...
  models.Subscription:
    properties:
      account_id:
        type: string
      created_at:
        type: string
      endpoint:
        type: string
      id:
        type: string
      kind:
        description: push | sms
        type: string
      metadata:
        type: string
    type: object
  models.Table:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      qr_code:
        type: string
    type: object
...
  models.TokenResponse:
    properties:
      account_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      role:
        example: customer
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Comprehensive restaurant management system with ordering, payments,
    inventory, and analytics
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Restaurant Management System API
  version: "2.0"
paths:
...
  /auth/signin:
    post:
      consumes:
      - application/json
      description: Authenticate with phone and password to receive a JWT token
      parameters:
      - description: Signin request
        in: body
        name: request
        required: true
        schema:
          properties:
            password:
              type: string
            phone:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Returns token, account_id, and role
          schema:
            $ref: '#/definitions/models.TokenResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Sign in to an existing account
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Create a new user account with phone and password
      parameters:
      - description: Signup request
        in: body
        name: request
        required: true
        schema:
          properties:
            password:
              type: string
            phone:
              type: string
            role:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Account created successfully
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Sign up a new user
      tags:
      - auth
...
  /kitchen/notifications/send:
    post:
      consumes:
      - application/json
      description: Send a push notification (staff/admin)
      parameters:
      - description: Notification payload
        in: body
        name: request
        required: true
        schema:
          properties:
            message:
              type: string
            subscription:
              $ref: '#/definitions/models.Subscription'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Notification sent
      security:
      - BearerAuth: []
      summary: Send notification
      tags:
      - notifications
...
  /notifications/{id}:
    delete:
      description: Remove notification subscription
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Unsubscribed
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Unsubscribe from notifications
      tags:
      - notifications
  /notifications/account/{account_id}:
    get:
      description: Get all notification subscriptions for an account
      parameters:
      - description: Account ID
        in: path
        name: account_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List notifications for account
      tags:
      - notifications
  /notifications/subscribe:
    post:
      consumes:
      - application/json
      description: Subscribe to push notifications
      parameters:
      - description: Subscription request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Subscription'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Subscription'
      security:
      - BearerAuth: []
      summary: Subscribe to notifications
      tags:
      - notifications
...
  /restaurant/{restaurant_id}/table/{table_id}/menu:
    get:
      description: Get menu for a specific restaurant table via QR code
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurant_id
        required: true
        type: string
      - description: Table ID
        in: path
        name: table_id
        required: true
        type: string
      - description: Language code
        in: query
        name: lang
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get QR menu
      tags:
      - menu
...
  /staff/branches/{branchId}/orders/{orderId}/assign-chef:
    post:
      consumes:
      - application/json
      description: Assign a chef to prepare an order
      parameters:
      - description: Branch ID
        in: path
        name: branchId
        required: true
        type: string
      - description: Order ID
        in: path
        name: orderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Assign chef to order
      tags:
      - staff
  /staff/branches/{branchId}/orders/{orderId}/split:
    post:
      consumes:
      - application/json
      description: Split an order into multiple orders
      parameters:
      - description: Branch ID
        in: path
        name: branchId
        required: true
        type: string
      - description: Order ID
        in: path
        name: orderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Split an order
      tags:
      - staff
  /staff/branches/{branchId}/orders/{orderId}/tip:
    post:
      consumes:
      - application/json
      description: Add a tip amount to an order
      parameters:
      - description: Branch ID
        in: path
        name: branchId
        required: true
        type: string
      - description: Order ID
        in: path
        name: orderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add tip to order
      tags:
      - staff
  /staff/branches/{branchId}/orders/merge:
    post:
      consumes:
      - application/json
      description: Merge multiple orders into one
      parameters:
      - description: Branch ID
        in: path
        name: branchId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Merge orders
      tags:
      - staff
  /staff/branches/{branchId}/tables/{tableId}/assign:
    post:
      consumes:
      - application/json
      description: Assign a waiter to a specific table
      parameters:
      - description: Branch ID
        in: path
        name: branchId
        required: true
        type: string
      - description: Table ID
        in: path
        name: tableId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Assign waiter to table
      tags:
      - staff
  /staff/branches/{branchId}/tables/{tableId}/state:
    patch:
      consumes:
      - application/json
      description: Update the state of a table (staff)
      parameters:
      - description: Branch ID
        in: path
        name: branchId
        required: true
        type: string
      - description: Table ID
        in: path
        name: tableId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update table state
      tags:
      - staff
  /tables:
    get:
      description: Get a list of all restaurant tables
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all tables
      tags:
      - tables
    post:
      consumes:
      - application/json
      description: Create a new restaurant table
      parameters:
      - description: Table request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Table'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Table'
      security:
      - BearerAuth: []
      summary: Create a new table
      tags:
      - tables
  /tables/{id}:
    get:
      description: Retrieve table details by ID
      parameters:
      - description: Table ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Table'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get table by ID
      tags:
      - tables
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"